pdf-viewer sin zoom utvider PDF'en fra midten, og ikke top-venste hjørne
